Goin to 0.1
Spawn points
Player colors
Bullet contact particles
Recharge animation
Recharge Icon
Game HUD
...

+++ Assets/Resources/Scripts/API/Damage.cs
@@ -1,9 +1,9 @@
 ﻿public class Damage
 {
-    public readonly float damage;
+    public readonly int damage;
     public readonly PlayerController owner;
 
-    public Damage(float damage, PlayerController owner)
+    public Damage(int damage, PlayerController owner)
     {
         this.damage = damage;
         this.owner = owner;
